requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
certifi==2024.8.30
defusedxml==0.7.1
diff-match-patch==20230430
distro==1.9.0
Django==5.0.3
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-filter==23.3
django-import-export==3.3.7
django-jazzmin==2.6.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-standardized-errors==0.13.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
flake8==6.1.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jiter==0.5.0
MarkupPy==1.14
mccabe==0.7.0
odfpy==1.4.1
openai==1.51.0
openpyxl==3.1.5
packaging==24.1
Pillow==10.1.0
pluggy==1.5.0
pycodestyle==2.11.1
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.1.0
PyJWT==2.8.0
pytest==7.4.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
redis==5.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tablib==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.7.0
xlrd==2.0.1
xlwt==1.3.0
